As we still have the problem of `bakcend.ls()` being slow on Artifactory (https://github.com/audeering/audbackend/issues/132) and the solution I use here does only work for legacy backends (as a anonymous user is not allowed to gather all folders on the root of the repo), I propose to speed up `audb` by adding manual code.

The execution time of `audb.versions("emodb")` is (averaged over 10 runs):

| main | this branch |
| --- | --- |
| 3.0 s | 1.2 s |

BTW, when not checking for the header file, but just for the folder, the execution time is 0.7 s. But this seems to risky to me, as the header file is usually our ground truth that the database is complete, compare https://github.com/audeering/audb/blob/e36d548c113b525b9e462302273be41c21b72bc6/audb/core/publish.py#L765-L775